feat: crud complete

+++ handlers/DeleteAluno.go
@@ -9,9 +9,18 @@ import (
 func DeleteAluno(ctx *gin.Context) {
 	var aluno models.Aluno
 	id := ctx.Params.ByName("id")
-	config.DB.Delete(&aluno, id)
 
+	config.DB.First(&aluno, id)
+	if aluno.ID == 0 {
+		ctx.JSON(404, gin.H{
+			"NotFound": "Aluno " + id + " Não encontrado",
+		})
+		return
+	}
+
+	config.DB.Delete(&aluno, id)
 	ctx.JSON(200, gin.H{
-		"Deleted": "Aluno " + id + " deletado",
+		"status":  "deleted",
+		"content": "Aluno " + id + " deletado",
 	})
 }
